refactor: simplify entry
